+++ api/IronCore/Audio/AudioTrack.cs
@@ -4,16 +4,16 @@ namespace Iron
 {
     public class AudioTrack
     {
-        internal AudioTrack(ulong id)
+        internal AudioTrack(uint id)
         {
             ID = id;
         }
         
-        public ulong ID { get; private set; }
+        public uint ID { get; private set; }
 
         public float Length => GetLength_Internal(ID);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        private static extern float GetLength_Internal(ulong audioTrackID);
+        private static extern float GetLength_Internal(uint audioTrackID);
     }
 }
