+++ api/IronCore/Entity/ComponentEngineCalls.cs
@@ -25,19 +25,19 @@ namespace Iron
             ((Component)GCHandle.FromIntPtr(componentPtr).Target)?.OnFixedUpdate();
         }
         
-        private static void ComponentOnCollisionEnter(IntPtr componentPtr, ulong otherEntityID)
+        private static void ComponentOnCollisionEnter(IntPtr componentPtr, uint otherEntityID)
         {
             Collision collision = new Collision(new Entity(otherEntityID));
             ((Component)GCHandle.FromIntPtr(componentPtr).Target)?.OnCollisionEnter(collision);
         }
         
-        private static void ComponentOnCollisionStay(IntPtr componentPtr, ulong otherEntityID)
+        private static void ComponentOnCollisionStay(IntPtr componentPtr, uint otherEntityID)
         {
             Collision collision = new Collision(new Entity(otherEntityID));
             ((Component)GCHandle.FromIntPtr(componentPtr).Target)?.OnCollisionStay(collision);
         }
         
-        private static void ComponentOnCollisionExit(IntPtr componentPtr, ulong otherEntityID)
+        private static void ComponentOnCollisionExit(IntPtr componentPtr, uint otherEntityID)
         {
             Collision collision = new Collision(new Entity(otherEntityID));
             ((Component)GCHandle.FromIntPtr(componentPtr).Target)?.OnCollisionExit(collision);
