+++ api/IronCore/Rendering/Sprite.cs
@@ -4,12 +4,12 @@ namespace Iron
 {
     public class Sprite
     {
-        internal Sprite(ulong id)
+        internal Sprite(uint id)
         {
             ID = id;
         }
         
-        public ulong ID { get; private set; }
+        public uint ID { get; private set; }
 
         public int Width => GetWidth_Internal(ID);
         public int Height => GetHeight_Internal(ID);
@@ -20,12 +20,12 @@ namespace Iron
         }
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        private static extern void SetAsSpriteSheet_Internal(ulong spriteID, int tileWidth, int tileHeight);
+        private static extern void SetAsSpriteSheet_Internal(uint spriteID, int tileWidth, int tileHeight);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        private static extern int GetWidth_Internal(ulong spriteID);
+        private static extern int GetWidth_Internal(uint spriteID);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        private static extern int GetHeight_Internal(ulong spriteID);
+        private static extern int GetHeight_Internal(uint spriteID);
     }
 }
